requirements.txt
albumentations==1.1.0
GDAL==3.0.2
hydra-core==1.3.2
matplotlib==3.5.1
numpy==1.21.6
omegaconf==2.0.6
opencv_contrib_python==3.4.2.17
opencv_python==4.8.0.76
opencv_python_headless==4.5.5.64
scikit_image==0.19.2
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
torch==1.13.0
tqdm==4.65.2
